% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{AllocateIdsRequest}
\alias{AllocateIdsRequest}
\title{Google Cloud Datastore API Objects 
Stores and queries data in Google Cloud Datastore.}
\usage{
AllocateIdsRequest(keys = NULL)
}
\arguments{
\item{keys}{A list of keys with incomplete key paths to allocate IDs for}
}
\value{
AllocateIdsRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:44:12
filename: /Users/mark/dev/R/autoGoogleAPI/googledatastorev1beta2.auto/R/datastore_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AllocateIdsRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AllocateIdsRequest functions: \code{\link{datasets.allocateIds}}
}
