% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{CommitRequest}
\alias{CommitRequest}
\title{CommitRequest Object}
\usage{
CommitRequest(ignoreReadOnly = NULL, mode = NULL, mutation = NULL,
  transaction = NULL)
}
\arguments{
\item{ignoreReadOnly}{No description}

\item{mode}{The type of commit to perform}

\item{mutation}{The mutation to perform}

\item{transaction}{The transaction identifier, returned by a call to beginTransaction}
}
\value{
CommitRequest object
}
\description{
CommitRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other CommitRequest functions: \code{\link{datasets.commit}}
}
