% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_functions.R
\name{datasets.rollback}
\alias{datasets.rollback}
\title{Roll back a transaction.}
\usage{
datasets.rollback(RollbackRequest, datasetId)
}
\arguments{
\item{RollbackRequest}{The \link{RollbackRequest} object to pass to this method}

\item{datasetId}{Identifies the dataset}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
\item https://www.googleapis.com/auth/userinfo.email
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore, https://www.googleapis.com/auth/userinfo.email)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/datastore/}{Google Documentation}

Other RollbackRequest functions: \code{\link{RollbackRequest}}
}
