% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{EntityResult}
\alias{EntityResult}
\title{EntityResult Object}
\usage{
EntityResult(entity = NULL)
}
\arguments{
\item{entity}{The resulting entity}
}
\value{
EntityResult object
}
\description{
EntityResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of fetching an entity from the datastore.
}

