% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{GqlQueryArg}
\alias{GqlQueryArg}
\title{GqlQueryArg Object}
\usage{
GqlQueryArg(cursor = NULL, name = NULL, value = NULL)
}
\arguments{
\item{cursor}{No description}

\item{name}{Must match regex '[A-Za-z_$][A-Za-z_$0-9]*'}

\item{value}{No description}
}
\value{
GqlQueryArg object
}
\description{
GqlQueryArg Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A binding argument for a GQL query.
}

