% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{KindExpression}
\alias{KindExpression}
\title{KindExpression Object}
\usage{
KindExpression(name = NULL)
}
\arguments{
\item{name}{The name of the kind}
}
\value{
KindExpression object
}
\description{
KindExpression Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A representation of a kind.
}

