% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{LookupResponse}
\alias{LookupResponse}
\title{LookupResponse Object}
\usage{
LookupResponse(deferred = NULL, found = NULL, header = NULL,
  missing = NULL)
}
\arguments{
\item{deferred}{A list of keys that were not looked up due to resource constraints}

\item{found}{Entities found}

\item{header}{No description}

\item{missing}{Entities not found, with only the key populated}
}
\value{
LookupResponse object
}
\description{
LookupResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

