% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{Mutation}
\alias{Mutation}
\title{Mutation Object}
\usage{
Mutation(delete = NULL, force = NULL, insert = NULL,
  insertAutoId = NULL, update = NULL, upsert = NULL)
}
\arguments{
\item{delete}{Keys of entities to delete}

\item{force}{Ignore a user specified read-only period}

\item{insert}{Entities to insert}

\item{insertAutoId}{Insert entities with a newly allocated ID}

\item{update}{Entities to update}

\item{upsert}{Entities to upsert}
}
\value{
Mutation object
}
\description{
Mutation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A set of changes to apply.
}
