% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{PartitionId}
\alias{PartitionId}
\title{PartitionId Object}
\usage{
PartitionId(datasetId = NULL, namespace = NULL)
}
\arguments{
\item{datasetId}{The dataset ID}

\item{namespace}{The namespace}
}
\value{
PartitionId object
}
\description{
PartitionId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An identifier for a particular subset of entities.Entities are partitioned into various subsets, each used by different datasets and different namespaces within a dataset and so forth.
}

