% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{Property}
\alias{Property}
\title{Property Object}
\usage{
Property(blobKeyValue = NULL, blobValue = NULL, booleanValue = NULL,
  dateTimeValue = NULL, doubleValue = NULL, entityValue = NULL,
  indexed = NULL, integerValue = NULL, keyValue = NULL,
  listValue = NULL, meaning = NULL, stringValue = NULL)
}
\arguments{
\item{blobKeyValue}{A blob key value}

\item{blobValue}{A blob value}

\item{booleanValue}{A boolean value}

\item{dateTimeValue}{A timestamp value}

\item{doubleValue}{A double value}

\item{entityValue}{An entity value}

\item{indexed}{If the value should be indexed}

\item{integerValue}{An integer value}

\item{keyValue}{A key value}

\item{listValue}{A list value}

\item{meaning}{The meaning field is reserved and should not be used}

\item{stringValue}{A UTF-8 encoded string value}
}
\value{
Property object
}
\description{
Property Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An entity property.
}

