% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{PropertyExpression}
\alias{PropertyExpression}
\title{PropertyExpression Object}
\usage{
PropertyExpression(aggregationFunction = NULL, property = NULL)
}
\arguments{
\item{aggregationFunction}{The aggregation function to apply to the property}

\item{property}{The property to project}
}
\value{
PropertyExpression object
}
\description{
PropertyExpression Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A representation of a property in a projection.
}

