% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{PropertyFilter}
\alias{PropertyFilter}
\title{PropertyFilter Object}
\usage{
PropertyFilter(operator = NULL, property = NULL, value = NULL)
}
\arguments{
\item{operator}{The operator to filter by}

\item{property}{The property to filter by}

\item{value}{The value to compare the property to}
}
\value{
PropertyFilter object
}
\description{
PropertyFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A filter on a specific property.
}

