% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{ResponseHeader}
\alias{ResponseHeader}
\title{ResponseHeader Object}
\usage{
ResponseHeader()
}
\value{
ResponseHeader object
}
\description{
ResponseHeader Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

