% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{RunQueryRequest}
\alias{RunQueryRequest}
\title{RunQueryRequest Object}
\usage{
RunQueryRequest(gqlQuery = NULL, partitionId = NULL, query = NULL,
  readOptions = NULL)
}
\arguments{
\item{gqlQuery}{The GQL query to run}

\item{partitionId}{Entities are partitioned into subsets, identified by a dataset (usually implicitly specified by the project) and namespace ID}

\item{query}{The query to run}

\item{readOptions}{The options for this query}
}
\value{
RunQueryRequest object
}
\description{
RunQueryRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RunQueryRequest functions: \code{\link{datasets.runQuery}}
}

